Netconf transactions synchronization VBD part
adding mechanism to sync transactions between VBD and GBP
this is used as a workaround for bug 
https://bugs.opendaylight.org/show_bug.cgi?id=7918
In order for this workaround to work VBD shares Reentrant lock to be 
used for synchronization of transactions between projects.
It is needed to use netconfSyncedWrite and netconfSyncedDelete 
from VbdNetconfTransaction for all write and delete operations to 
netconf device or to implement project specific write/delete 
methods which will use the same lock.

Change-Id: I479bd140b6e18ba4c58640f920c4b60b19365537
Signed-off-by: Michal Cmarada <mcmarada@cisco.com>