Update patch set 5
Patch Set 5:

Yes, it does mean we will not be compatible with 17.01. And there is so far no way of being compatible with both 17.01 and 17.04 as the previous contains old ( rev 2016-12-14) and the later contains new (rev 17-03-15) yangs only. That being said, the new yangs enable some of the very required functionalities.

Patch-set: 5
Label: Verified=0