Add Export version for lisp address
ietf-lisp-address yang exists in both this project and LFM. LFM has
version 1.6.0 and VBD has version 1.2.0. The problem here is, if vbd
feature is installed before LFM feature, then LFM can't export the
package and because LFM require the packages to have version >= 1.6.0,
this patch export VBD's lisp address packages to have version 1.6.0
too making it possible to install the features in any order desired.
This patch is a quick fix.

Change-Id: I73a53e1935847b2fac8f5db11a68a06d3557d27e
Signed-off-by: Shakib Ahmed <sheikahm@cisco.com>